+++ monkey/infection_monkey/exploit/mssqlexec.py
@@ -11,7 +11,7 @@ from infection_monkey.exploit.tools.http_tools import HTTPTools
 from infection_monkey.exploit.tools.helpers import get_monkey_dest_path, get_target_monkey, \
     build_monkey_commandline, get_monkey_depth
 from infection_monkey.model import DROPPER_ARG
-from infection_monkey.utils import get_monkey_dir_path
+from infection_monkey.utils.monkey_dir import get_monkey_dir_path
 
 LOG = logging.getLogger(__name__)
