+++ monkey/infection_monkey/exploit/sambacry_monkey_runner/build.sh
