+++ monkey/infection_monkey/exploit/sambacry_monkey_runner/sc_monkey_runner.c
