+++ monkey/infection_monkey/utils/windows/new_user.py
